refactor: remove cache mechanism and old code
Not using cache anymore because it does not have much impact on
performance improvement. Using the custom definition of
s:HL within s:set_syntax_highlighting function has much greater
performance by reducing a lot of overhead of if-branch
conditions because s:HL function is called extensively.

The performance is close to the solution for issue #28, which was
for version 0.7. This current improvement is for the latest version.

see: #28