Set cterm font features in MODE_GUI_COLOR.
If the user is using a terminal but has termguicolors enabled, vim will
use the guifg and guibg for colors, but cterm for features. PaperColor
detects termguicolors and moves forward in MODE_GUI_COLOR, but doesn't
set cterm, so the 24bit color terminal user doesn't get the benefit of
bold and italic features.