Add option to skip file-type specific colors
When turned on, this should fix the problem with the color scheme
pollitung filetype-specific highlight groups that are by default linked
to more generic groups at the cost of disabling all filetype-specific
colouring from the theme.

It can be switched on by setting g:only_generic_hl to a truthy value
before loading the colourscheme.